+++ staramr/databases/resistance/ARGDrugTable.py
@@ -24,7 +24,9 @@ class ARGDrugTable:
         self._file = file
 
         if file is not None:
-            self._data = pd.read_csv(file, sep='\t', dtype=self.DTYPES)
+            # "None" is recognized as a NA/NaN string in pandas 2.
+            # However, in pandas < 2, "None" is not a default NA value, so we must be explicit.
+            self._data = pd.read_csv(file, sep='\t', dtype=self.DTYPES, na_values="None")
 
     def get_resistance_table_info(self):
         """
@@ -41,4 +43,10 @@ class ARGDrugTable:
         :param drug: The drug string.
         :return: The drug string with correct separators/spacing.
         """
-        return ', '.join(drug.split(','))
+
+        if type(drug) is str:
+            result = ', '.join(drug.split(','))
+        else:
+            result = drug
+
+        return result